@@ -4,22 +4,31 @@
   "author": "BCsabaEngine <csaba@balazs-magyar.hu>",
   "version": "0.9.6",
   "license": "ISC",
-  "exports": "./RF24MeshSerialNode.js",
-  "main": "RF24MeshSerialNode.js",
+  "exports": "./dist/RF24MeshSerialNode.js",
+  "main": "./dist/RF24MeshSerialNode.js",
+  "types": "./dist/RF24MeshSerialNode.d.ts",
   "engines": {
-    "node": ">=12",
-    "npm": ">=7"
+    "node": ">=18",
+    "npm": ">=9"
   },
+  "files": [
+    "**/*.js",
+    "**/*.d.ts"
+  ],
   "repository": {
     "type": "git",
-    "url": "https://github.com/SegmentController/RF24MeshSerialNode"
+    "url": "git+https://github.com/SegmentController/RF24MeshSerialNode.git"
   },
   "bugs": {
     "url": "https://github.com/SegmentController/RF24MeshSerialNode/issues"
   },
   "homepage": "https://github.com/SegmentController/RF24MeshSerialNode",
   "scripts": {
-    "npm-publish": "npm version patch && npm publish && git push"
+    "dev": "nodemon examples/simpleNode.ts",
+    "clean": "tsc --build --clean",
+    "build": "tsc --build --clean && tsc --build --force",
+    "npm-publish-patch": "npm run build && npm version patch && npm publish && npm run clean && git push",
+    "npm-publish-minor": "npm run build && npm version minor && npm publish && npm run clean && git push"
   },
   "keywords": [
     "nodejs",
@@ -29,6 +38,12 @@
     "serialbridge"
   ],
   "dependencies": {
-    "serialport": "^10.4.0"
+    "serialport": "^12.0.0"
+  },
+  "devDependencies": {
+    "@types/node": "20.9.1",
+    "nodemon": "^3.0.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2"
   }
 }
